driver: src: io_backend: Work around to fix infinite loop
Compiling the driver with -ftree-pre and -fthread-jumps is causing and
infinite while loop. We have observed that replacing logic with the for
loop avoids this behavior. This is a work around so we can continue to
build driver with complete set of -O2 optimizations. Further debugging
of this issue will be required.

Signed-off-by: Gregory Williams <gregory.williams@amd.com>

Sankarji Gopalakrishnan